Sync package structure and cleanup dependencies
- Move LaunchTime.kt to com.example.android.performance.vitals to match package declaration.
- Remove redundant explicit dependencies (compose-runtime, lifecycle, and coroutines) as they are provided transitively by activity-compose.
